Download a data series from the IMF

Function to request data from a database through the IMF API endpoint.

imf_data(database_id, parameters, start_year, end_year ...)

Arguments

database_id

character string. Database ID for database from which you would like to request data. Can be found using imf_databases.

parameters

list of data frames providing input parameters for your API request. Retrieve list of all possible input parameters using imf_parameters and filter each data frame in the list to reduce it to the inputs you want.

start_year

integer four-digit year. Earliest year for which you would like to request data.

end_year

integer four-digit year. Latest year for which you would like to request data.

print_url

logical. Whether to print the URL used in the API call. Can be useful for debugging.

times

numeric. Maximum number of requests to attempt.

accounting_entry, activity, adjustment, age, classification, cofog_function, commodity, comp_method, composite_breakdown, counterpart_area, counterpart_sector, currency_denom, cust_breakdown, disability_status, education_lev, expenditure, financial_institution, flow_stock_entry, freq, functional_cat, gfs_sto, income_wealth_quantile, indicator, instr_asset, instrument_and_assets_classification, int_acc_item, maturity, occupation, prices, product, ref_area, ref_sector, reporting_type, series, sex, sto, summary_statistics, survey, transformation, type, unit_measure, urbanisation, valuation

character vector. Use imf_parameters to identify which parameters to use for requests from a given database and to see all valid input codes for each parameter.

Value

Returns a tidy data frame with the data series.

Details

Only the database_id argument is strictly required; all other arguments are optional. If you provide a database_id without any other arguments, the function will attempt to download the entire database. However, many databases available through the API are too large to download in their entirety, and your request will fail. Additional arguments to the function act as filter parameters to reduce the size of the returned dataset. For instance, supplying c("A","M") as the freq argument will return all database observations of annual or monthly frequency, while excluding all observations of quarterly frequency.

There are two ways to supply parameters for your API request. The optimal way is to retrieve a list of data frames using imf_parameters, filter each data frame to retain only the parameters you want, and then supply the modified list object to imf_data as its parameters argument. However, users who are not comfortable modifying data frames in a nested list may find it easier to instead supply one or more character vectors as arguments, as in the example in the previous paragraph. (There are a total of 44 possible parameters for making request from various databases through the API, and each parameter uses unique input codes, which is why the parameters list method simplifies things!) These two methods for specifying parameters may not be combined. Only database_id, start_year, end_year, print_url, and times arguments may be used in combination with a parameters list object; any other arguments will be ignored (and a warning thrown).

Examples

# Retrieve "Assets (with Fund Record), National Currency" series from the
# Balance of Payments database using the list object method
params <- imf_parameters(database_id = 'BOP')
params$indicator <- filter(params$indicator,
                           description == 'Assets (with Fund Record), National Currency')
#> Error in filter(params$indicator, description == "Assets (with Fund Record), National Currency"): object 'description' not found
df <- imf_data(database_id = 'BOP', parameters = params)
#> Request failed [404]. Retrying in 1.1 seconds...
#> Request failed [404]. Retrying in 3.6 seconds...
#> Error: data.imf.org appears to be down.

# Retrieve "Assets (with Fund Record), National Currency" series from the
# Balance of Payments database using the character vector method
params <- imf_parameters(database_id = 'BOP')
indicator_code <- filter(params$indicator,
                           description == 'Assets (with Fund Record), National Currency')$input_code
#> Error in filter(params$indicator, description == "Assets (with Fund Record), National Currency"): object 'description' not found
df <- imf_data(database_id = 'BOP', indicator = indicator_code)
#> Error in map(1:length(name_vector), function(x) {    if (name_vector[x] %in% vector_vars[supplied_vectors]) {        df <- filter(data_dimensions[[x]], input_code %in% eval(parse(text = name_vector[x])))        if (any(!(eval(parse(text = name_vector[x])) %in% df$input_code))) {            warning(paste0(sum(!(eval(parse(text = name_vector[x])) %in%                 df$input_code)), " invalid user-supplied input code(s) for ",                 name_vector[x], " parameter will be ignored.\nUse imf_parameters('",                 database_id, "')$", name_vector[x], " to see all valid input codes for this parameter.\nNote that codes are case sensitive.",                 immediate. = T))        }        df <- if (nrow(df) == nrow(data_dimensions[[x]])) {            data.frame(input_code = c(), description = c())        }        else {            df        }        return(df)    }    else {        return(data.frame(input_code = c(), description = c()))    }}): ℹ In index: 3.
#> Caused by error in `filter()`:
#> ℹ In argument: `input_code %in% eval(parse(text = name_vector[x]))`.
#> Caused by error:
#> ! object 'indicator_code' not found